# Use a 128 sample buffer in memSeries