--- a/querier.go
+++ b/querier.go
@@ -677,6 +677,8 @@ type SeriesIterator interface {
 	Next() bool
 	// Err returns the current error.
 	Err() error
+	// Close releases any resources held by the iterator.
+	Close()
 }
 
 // chainedSeries implements a series for a list of time-sorted series.
@@ -716,8 +718,10 @@ func (it *chainedSeriesIterator) Seek(t int64) bool {
 	for i, s := range it.series[it.i:] {
 		cur := s.Iterator()
 		if !cur.Seek(t) {
+			cur.Close()
 			continue
 		}
+		it.cur.Close()
 		it.cur = cur
 		it.i += i
 		return true
@@ -730,16 +734,23 @@ func (it *chainedSeriesIterator) Next() bool {
 		return true
 	}
 	if err := it.cur.Err(); err != nil {
+		it.cur.Close()
 		return false
 	}
 	if it.i == len(it.series)-1 {
+		it.cur.Close()
 		return false
 	}
 
 	it.i++
+	it.cur.Close()
 	it.cur = it.series[it.i].Iterator()
 
-	return it.Next()
+	if !it.Next() {
+		it.cur.Close()
+		return false
+	}
+	return true
 }
 
 func (it *chainedSeriesIterator) At() (t int64, v float64) {
@@ -748,6 +759,10 @@ func (it *chainedSeriesIterator) At() (t int64, v float64) {
 
 func (it *chainedSeriesIterator) Err() error {
 	return it.cur.Err()
+}
+
+func (it *chainedSeriesIterator) Close() {
+	it.cur.Close()
 }
 
 // chunkSeriesIterator implements a series iterator on top
@@ -797,6 +812,7 @@ func (it *chunkSeriesIterator) Seek(t int64) (ok bool) {
 		}
 	}
 
+	it.cur.Close()
 	it.cur = it.chunks[it.i].Chunk.Iterator()
 	if len(it.intervals) > 0 {
 		it.cur = &deletedIterator{it: it.cur, intervals: it.intervals}
@@ -808,6 +824,7 @@ func (it *chunkSeriesIterator) Seek(t int64) (ok bool) {
 			return true
 		}
 	}
+	it.cur.Close()
 	return false
 }
 
@@ -825,14 +842,20 @@ func (it *chunkSeriesIterator) Next() bool {
 			}
 			t, _ = it.At()
 
-			return t <= it.maxt
+			if t > it.maxt {
+				it.cur.Close()
+				return false
+			}
+			return true
 		}
 		if t > it.maxt {
+			it.cur.Close()
 			return false
 		}
 		return true
 	}
 	if err := it.cur.Err(); err != nil {
+		it.cur.Close()
 		return false
 	}
 	if it.i == len(it.chunks)-1 {
@@ -840,6 +863,7 @@ func (it *chunkSeriesIterator) Next() bool {
 	}
 
 	it.i++
+	it.cur.Close()
 	it.cur = it.chunks[it.i].Chunk.Iterator()
 	if len(it.intervals) > 0 {
 		it.cur = &deletedIterator{it: it.cur, intervals: it.intervals}
@@ -850,6 +874,10 @@ func (it *chunkSeriesIterator) Next() bool {
 
 func (it *chunkSeriesIterator) Err() error {
 	return it.cur.Err()
+}
+
+func (it *chunkSeriesIterator) Close() {
+	it.cur.Close()
 }
 
 // deletedIterator wraps an Iterator and makes sure any deleted metrics are not
@@ -890,6 +918,10 @@ Outer:
 
 func (it *deletedIterator) Err() error {
 	return it.it.Err()
+}
+
+func (it *deletedIterator) Close() {
+	it.it.Close()
 }
 
 type mockSeriesSet struct {