This is intended as a performance improvement for instant and short range queries, i.e. most rule evals. It expands the `memSeries` buffer from 4 to 128 samples, so it covers the whole last chunk, avoiding decoding the whole chunk for the samples most likely to be queried by rule evaluation.

It increases memory utilization by 2 KB per head series, improving instant and short range matrix query times by 20%+. Memory utilization could be reduced, either by lowering the buffer size (not very useful, as the iterator would still have to decode the part of the chunk not covered by the buffer) or by selectively using larger buffers for queried series only.

[Before](https://github.com/prometheus/tsdb/files/2164228/before.128.txt) and [after](https://github.com/prometheus/tsdb/files/2164229/after.128.txt) benchmark results and [comparison](https://github.com/prometheus/tsdb/files/2164230/benchcmp.128.txt) are attached.

The complete change (including the Prometheus side of things) is in [free/prometheus](https://github.com/free/prometheus/commit/f83000805fd473ce501077716f27144e7d2ca535).