@fabxc 2018-07-04
> It increases memory utilization by 2 KB per head series

By past measurement that's the current overall cost of one head series, i.e. this would double Prometheus's memory usage. Did you try this out in a live Prometheus server?