@free 2018-07-05
> By past measurement that's the current overall cost of one head series, i.e. this would double Prometheus's memory usage.

It is actually exactly a 2 KB per series increase in memory usage. (The iterator buffers are pooled and only one iterator is live at a time for each query, so they really don't matter.) So the memory increase is pretty easy to estimate for every Prometheus instance.

> Did you try this out in a live Prometheus server?

Yes I did, in fact. One with ~50K series. (It also includes all my other proposed Prometheus performance improvements (prometheus/prometheus#4248, prometheus/prometheus#4271, prometheus/prometheus#4315 and prometheus/tsdb#351) so it's not exactly an apples to apples comparison, but also shouldn't affect steady state memory usage. Regardless, on this particular server (with a limited number of series, relatively expensive recorded rules and very little user queries) steady state memory usage is higher by an expected ~100 MB.

Here's the memory profile, for reference: [pprof.prometheus.alloc_objects.alloc_space.inuse_objects.inuse_space.022.pb.gz](https://github.com/prometheus/tsdb/files/2165565/pprof.prometheus.alloc_objects.alloc_space.inuse_objects.inuse_space.022.pb.gz). It shows 115 MB allocated by `newMemSeries` (new series creation, in 2.25 KB chunks), compared to 189 MB allocated from below `(*memSeries).append` (appending to existing series), so about a 60% increase in memory used by TSDB. (I use a 5s scrape interval and 10s eval interval, the percentage will likely be higher with e.g. 1m scrape intervals.)

If this turns out to be unacceptable, one could instead stick with the 4 sample buffer by default and switch to a dynamically allocated slice for queried series only. On the one hand that would limit the memory usage increase by buffering fewer series (those for which iterator speed matters) and on the other hand might use smaller buffer sizes for short-lived series. There would be a small increase in mutex contentions the first time a series is queried (because the samples need to be decoded into the buffer), but that would only be an issue (if at all) for one-off queries that touch previously unqueried series. (Rule eval would presumably do the locking either when the series is created, and thus the chunk is almost empty, or at startup.)