@krasi-georgiev 2018-12-17
@free would you have time to resolve the conflict and I will look into this.

I am running some benchmarks as part of another PR so have an isolated machine to compare the results.
Can also run a proper Prombench test as well with these changes.