@free 2018-12-18
Conflicts resolved, tests are passing. I haven't had the time to build and test a Prometheus instance with this latest version, but I wouldn't expect any issues (or differences from my earlier test).

Like I said in my earlier comments, this change could be adjusted to dynamically allocate a buffer only for series that are being queried (as opposed to all series). And the buffer size could be configurable, e.g. if you know you're scraping every 10 seconds and your rules use maximum ranges of 5 minutes, then you could use a buffer size of 30 + epsilon instead of 128, as you'd only be interested in the last 30 samples when doing rule evaluation.

Oh, and finally, as part of the merge I've had to revert @brian-brazil's #444, because `sampleBuf` is now a circular buffer and so the computation of the index may not be worth the memory savings (especially seeing as how this change trades memory usage for CPU anyway).