@brian-brazil 2018-12-18
> Like I said in my earlier comments, this change could be adjusted to dynamically allocate a buffer only for series that are being queried (as opposed to all series)

This is still potentially all series, as it's easy for a user to request all of them and cause a sudden bump in memory usage.

For a Prometheus with 10M series that's an extra 4GB of RAM (have to double the usage due to GC). I don't think that's a good tradeoff.