@free 2018-12-18
> For a Prometheus with 10M series that's an extra 4GB of RAM (have to double the usage due to GC). I don't think that's a good tradeoff.

4GB on its own looks like a big, scary number. But when considered in relation to the baseline memory usage of a Prometheus with 10M series, I doubt it looks all that bad. I run a Prometheus with only 130K series and a baseline memory usage of ~2.5GB, so an extra 2KB per series is merely 10% overhead. The percentage may be larger for a Prometheus instance with 10M series, but I for one would gladly trade 10% RAM for 20% CPU. And if I had to put money on it, I'd guess most Prometheus deployments are closer to 130K than 10M series.

Even better, it would be great to have a knob I could turn where I am free to choose between 20% more RAM or 20% more CPU usage. Or anything in-between. Which is fully doable.

> This is still potentially all series, as it's easy for a user to request all of them and cause a sudden bump in memory usage.

It is just as easy for a user to issue 10 parallel queries going back 1 year or covering a few thousand series. I'm sure you can easily get a Prometheus with 10M series (or 130K series, for that matter) to blow up that way, no 2KB/series overhead needed. And you don't need some perfect storm for that to happen, just some clueless user selecting a 1 year range on a Grafana dashboard.

Plus, if this is the main concern, it should be relatively easy to either add a counter to each series and only allocate the buffer once the series was queried more than N times (with some exponential decay thrown in). Or limit the buffering to series queried by rule evaluation. Or a few other ways.