@brian-brazil 2018-12-18
> And if I had to put money on it, I'd guess most Prometheus deployments are closer to 130K than 10M series.

We have many users with excessive cardinality, and we hear a lot from them.

> It is just as easy for a user to issue 10 parallel queries going back 1 year or covering a few thousand series. 

As of 2.5.0, there's protections for that. That's also memory usage only for the lifetime of the query, not the lifetime of the series.

> Or limit the buffering to series queried by rule evaluation. 

Some users do put big queries in rule evaluation. I don't think it's safe operationally to have a varying value for this.