@krasi-georgiev 2018-12-18
the initial bench results show some gains and losses in performance.

```
benchcmp -mag -best old128SampleBuffer.txt new128SampleBuffer.txt 
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_ten,steps=1-16                                                                              121834         100264         -17.70%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-16                                                             1525009607     1280496030     -16.03%
BenchmarkRangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1-16                                      143303         169895         +18.56%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-16                                                          2049532384     2427188884     +18.43%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-16                                                  319685         270065         -15.52%
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-16                                    518656298      610938446      +17.79%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=100-16              967917         1134796        +17.24%
BenchmarkRangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=1000-16                                       125753845      107611840      -14.43%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-16                                                8085431        9422173        +16.53%
BenchmarkRangeQuery/expr=sum_without_(l)(h_one),steps=100-16                                                           1416127        1220579        -13.81%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-16                                                           94911985       109603131      +15.48%
BenchmarkRangeQuery/expr=abs(a_hundred),steps=1-16                                                                     1233212        1070040        -13.23%
BenchmarkRangeQuery/expr=sum_by_(l)(h_one),steps=10-16                                                                 130891         150399         +14.90%
BenchmarkRangeQuery/expr=abs(a_one),steps=1-16                                                                         32564          37414          +14.89%
BenchmarkRangeQuery/expr=rate(a_one[1m])_+_rate(b_one[1m]),steps=10-16                                                 81303          70909          -12.78%
BenchmarkRangeQuery/expr=-a_one,steps=1-16                                                                             35849          31373          -12.49%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-16                                                 342533         300642         -12.23%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-16                                                                   1054326        934382         -11.38%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1-16                                                   47953          42934          -10.47%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-16                                            28271676       25552649       -9.62%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-16                                                                 2043487        1847455        -9.59%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-16                                                 1840960        2030539        +10.30%
BenchmarkRangeQuery/expr=sum_without_(l)(h_one),steps=10-16                                                            262000         237881         -9.21%
BenchmarkRangeQuery/expr=sum_by_(le)(h_ten),steps=100-16                                                               3617587        3982531        +10.09%
BenchmarkRangeQuery/expr=-a_one,steps=100-16                                                                           44151          40182          -8.99%
BenchmarkRangeQuery/expr=abs(a_one),steps=10-16                                                                        35829          39178          +9.35%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-16                                               773132         707293         -8.52%
BenchmarkRangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-16                                       5921697        6466325        +9.20%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-16                                                   740113         805824         +8.88%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-16                                                                    148185         136132         -8.13%
BenchmarkRangeQuery/expr=sum_without_(le)(h_ten),steps=1-16                                                            1066798        980667         -8.07%
BenchmarkRangeQuery/expr=a_one,steps=1000-16                                                                           124525         115186         -7.50%
BenchmarkRangeQuery/expr=-a_ten,steps=10-16                                                                            109560         101547         -7.31%
BenchmarkRangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=10-16               293979         316468         +7.65%
BenchmarkRangeQuery/expr=sum_without_(le)(h_ten),steps=10-16                                                           1255500        1170190        -6.79%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-16                                                                1069964        999694         -6.57%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-16                                                                    34317          36665          +6.84%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-16                                                                   36003          38342          +6.50%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-16                                                            11154411       11873247       +6.44%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-16                                                                     55085          51814          -5.94%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-16                                                  144936         137033         -5.45%
BenchmarkRangeQuery/expr=sum(a_one),steps=100-16                                                                       103514         97969          -5.36%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-16                                                                      47120          44797          -4.93%
BenchmarkRangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1000-16                                   675299         710159         +5.16%
BenchmarkRangeQuery/expr=sum_by_(l)(h_one),steps=100-16                                                                342576         360116         +5.12%

benchmark                                                                                                              old bytes     new bytes     delta
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-16                                    10640763      12621088      +18.61%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-16                                                             3183988       2804738       -11.91%
BenchmarkRangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-16                                                      10581144      11118767      +5.08%
BenchmarkRangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-16                                                     60444334      58485049      -3.24%
```